Move AppearanceSettingsFragment to a new file
